the cli is shaping up
