+++ web10-cli/bin/commands.js
@@ -1,0 +1,7 @@
+#! /usr/bin/env node
+const args = process.argv.slice(2);
+console.log("")
+console.log("  commands - a help menu of possible web10 cli commands")
+console.log("  create - create a web10 app")
+console.log("")
+process.exit(0); //no errors occurred
