+++ web10-cli/package.json
@@ -4,7 +4,7 @@
   "description": "a command line interface with utilities made by web10 inc. in order to make web10 easier to use.",
   "main": "index.js",
   "bin": {
-    "help":"bin/help.js",
+    "commands":"bin/commands.js",
     "create": "bin/create.js"
   },
   "scripts": {
